requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
email-validator==2.2.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
python-jose==3.5.0
alembic==1.16.4
google-auth-oauthlib==1.2.2
google-api-python-client==2.166.0
